extract download link wrangler to separate module
